requirements.txt
faiss-gpu==1.7.2
pymysql-pool==0.4.6
numpy==1.23.5
pymysql-pool==0.4.6
fastapi==0.115.8
uvicorn==0.34.0
gunicorn==23.0.0
oracledb==2.5.1
requests==2.32.4